+++ Tving-Clone-Project/Tving-Clone-Project/App/SceneDelegate.swift
@@ -14,8 +14,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: windowScene)
-//        self.window?.rootViewController = LoginViewController(viewModel: LoginViewModel())
-        self.window?.rootViewController = TabBarController()
+        self.window?.rootViewController = LoginViewController(viewModel: LoginViewModel())
+//        self.window?.rootViewController = TabBarController()
         self.window?.makeKeyAndVisible()
     }
 }
